refactor(admin): move to package of files
